[project @ 2004-05-23 20:49:22 by panne] Added Sven Eric Panitz' HOpenGL tutorial. It contains only small superficial changes compared to the version kindly provided by my namesake, the contents of the tutorial itself haven't changed yet. No HTML output is generated yet, because tth is not freely available and using latex2html doesn't work yet. The examples don't compile, either, some simple API tweaking is still needed.
Note that currently one needs Sofware AG's quip for building the documentation
and extracting the examples. Perhaps this will change in the future when e.g.
DocBook is used, but for now we simply leave out the doc directory from the
build of the GLUT package. Perhaps even some generated files will be checked
into the repository, which is not nice, but workable for a transitional period.